rename vitest snapshot serializer
