+++ packages/internals/src/__tests__/engine-commands/getConfig.test.ts
@@ -1,4 +1,4 @@
-import { serialize } from '@prisma/get-platform/src/test-utils/vitestSnapshotSerializer'
+import { serialize } from '@prisma/get-platform/src/test-utils/vitest-snapshot-serializer'
 import { describe, expect, test } from 'vitest'
 
 import { getConfig, isRustPanic } from '../..'
